chore: moving the library back to the root dir
